add encode.go
