+++ message.go
@@ -1,11 +1,5 @@
 package cat
 
-import "bytes"
-
-type Encodable interface {
-	Encode(*bytes.Buffer) Error
-}
-
 type Message interface {
 	Meta
 	Encodable
